requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
python-jose
python-jose[cryptography]
passlib
pytz
redis
python-multipart
bcrypt